dispatcher: Use int32 for atomic operations to prevent crash on ARM 32bit systems
See https://github.com/golang/go/issues/599

+++ dispatch/dispatch_types.go
@@ -46,11 +46,11 @@ type Dispatcher struct {
 	outbound sync.Pool
 
 	// MaxWorkers defines max worker ceiling
-	maxWorkers int64
+	maxWorkers int32
 
 	// Atomic values -----------------------
 	// Worker counter
-	count int64
+	count int32
 	// Dispatch status
 	running uint32
